Move aggregator to internal/search/run
Will reorganize the internal/search/run package in the near future, but
for now it is a dumping ground as I rip things out of graphqlbackend.
